pyproject.toml
[project]
name = "llumdocs"
version = "0.1.0"
description = "LlumDocs: plataforma per transformar i entendre documents amb IA"
authors = [
  { name = "Marc Maceira" },
]
requires-python = ">=3.12"
dependencies = [
  "litellm>=1.0.0",
  "openai>=1.0.0",
  "gradio",
  "typer",
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv",
  "httpx",
  "numpy",
  "Pillow",
  "docling",
  "PyMuPDF",
  "pdf2image",
  "pytesseract",
  "rapidocr-onnxruntime",
]

[project.scripts]
llumdocs-api = "llumdocs.api.app:main"
llumdocs-ui = "llumdocs.ui.main:main"

[project.optional-dependencies]
email = [
  "torch>=2.5.0",
  "transformers>=4.45.0",
]
dev = [
  "pre-commit",
  "ruff",
  "black",
  "pytest",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

# 👇 NEW: tell setuptools which packages to include/exclude
[tool.setuptools.packages.find]
where = ["."]
include = ["llumdocs*"]
exclude = ["setup*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
extend-select = ["E", "F", "I", "B"]

[tool.ruff.lint.isort]
known-first-party = ["llumdocs"]
combine-as-imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
filterwarnings = [
    "ignore::UserWarning:pydantic.main",
    "ignore::DeprecationWarning:litellm._service_logger",
]
